[skeleton] postprocess: refactor the default plugin
