+++ projects/skeleton/postprocess/default/parsing/__init__.py
@@ -1,0 +1,6 @@
+"""Parsing package for Skeleton Caliper plugin."""
+
+from .kpis import SkeletonKpiHandler
+from .parsers import SkeletonParser
+
+__all__ = ["SkeletonParser", "SkeletonKpiHandler"]
