+++ projects/skeleton/postprocess/default/plotting/__init__.py
@@ -1,0 +1,6 @@
+"""Plotting package for Skeleton Caliper plugin."""
+
+from .summary_table import SummaryTablePlot
+from .throughput_chart import ThroughputChartPlot
+
+__all__ = ["SummaryTablePlot", "ThroughputChartPlot"]
